refactor: move set-pieces from node-graph/ to network/
Separates content/generation (js/core/network/) from runtime (js/core/node-graph/).
set-pieces.js and its test move to the new directory. All imports updated.
523 tests pass, no behavior change.
