+++ data/networks/corporate-foothold.js
@@ -7,7 +7,7 @@
  * No ICE.
  */
 
-import { instantiate, SET_PIECES } from "../../js/core/node-graph/set-pieces.js";
+import { instantiate, SET_PIECES } from "../../js/core/network/set-pieces.js";
 import {
   createGateway, createRouter, createWAN, createGameNode,
 } from "../../js/core/node-graph/game-types.js";
