+++ data/networks/research-station.js
@@ -7,7 +7,7 @@
  * Set-pieces: deadmanCircuit, combinationLock, encryptedVault, tamperDetect.
  */
 
-import { instantiate, SET_PIECES } from "../../js/core/node-graph/set-pieces.js";
+import { instantiate, SET_PIECES } from "../../js/core/network/set-pieces.js";
 import {
   createGateway, createRouter, createFileserver, createWAN, createGameNode,
 } from "../../js/core/node-graph/game-types.js";
